Track file descriptors instead of streams.
I'm about to implement read/write files which are easier to track as file
descriptors (one for read, one for write) so it makes sense to track read-only
and write-only files this way. Conveniently a file descriptor is an int and
FUSE gives us a 64-bit handle so we can pack two descriptors into the handle.
The low 32 bits are used for the read descriptor and the upper 32 bits for the
write descriptor. 0 indicates that the file is not open for this mode, which is
fine because the descriptor could never be referring to stdin.