Tests: fix: Workaround for race in write test.
When writing to the output file in this test there is a race condition with
regard to reading the input back in. Because the write and read are done in
separate processes there are no guarantees about the ordering of these
operations. This commit introduces a sleep which is a bit of a hack, but works.